requirements.txt
fastapi
uvicorn
motor
psycopg2-binary
neo4j
pandas
openpyxl
python-dotenv